@@ -21,7 +21,7 @@ import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.Dimension;
+import androidx.annotation.Dimension;
 import android.util.AttributeSet;
 import android.view.View;
 
