@@ -16,8 +16,8 @@
 
 package io.github.marktony.espresso.data.source.remote;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.List;
 
