@@ -20,7 +20,7 @@ import android.view.View;
 
 /**
  * Created by lizhaotailang on 2017/2/11.
- * An interface for {@link android.support.v7.widget.RecyclerView} item.
+ * An interface for {@link androidx.recyclerview.widget.RecyclerView} item.
  */
 
 public interface OnRecyclerViewItemClickListener {
