@@ -18,9 +18,9 @@ package io.github.marktony.espresso.mvp.packagedetails;
 
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.annotation.Nullable;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
 
 import io.github.marktony.espresso.R;
 import io.github.marktony.espresso.data.source.local.PackagesLocalDataSource;
