@@ -53,14 +53,12 @@ public class DecodeHandler extends Handler {
 		if (!running) {
 			return;
 		}
-		switch (message.what) {
-		case R.id.decode:
+		int what = message.what;
+		if (what == R.id.decode) {
 			decode((byte[]) message.obj, message.arg1, message.arg2);
-			break;
-		case R.id.quit:
+		} else if (what == R.id.quit) {
 			running = false;
 			Looper.myLooper().quit();
-			break;
 		}
 	}
 
