@@ -59,25 +59,21 @@ public class CaptureActivityHandler extends Handler {
 
 	@Override
 	public void handleMessage(Message message) {
-		switch (message.what) {
-		case R.id.restart_preview:
+		int what = message.what;
+		if (what == R.id.restart_preview) {
 			restartPreviewAndDecode();
-			break;
-		case R.id.decode_succeeded:
+		} else if (what == R.id.decode_succeeded) {
 			state = State.SUCCESS;
 			Bundle bundle = message.getData();
 			activity.handleDecode((Result) message.obj, bundle);
-			break;
-		case R.id.decode_failed:
+		} else if (what == R.id.decode_failed) {
 			// We're decoding as fast as possible, so when one decode fails,
 			// start another.
 			state = State.PREVIEW;
 			cameraManager.requestPreviewFrame(decodeThread.getHandler(), R.id.decode);
-			break;
-		case R.id.return_scan_result:
+		} else if (what == R.id.return_scan_result) {
 			activity.setResult(Activity.RESULT_OK, (Intent) message.obj);
 			activity.finish();
-			break;
 		}
 	}
 
